Cleanup project structure
