+++ blockycraft/Assets/Behaviours/Chunk.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using Assets.Scripts.World.Chunk;
+using UnityEngine;
 
 public sealed class Chunk
 {
