+++ blockycraft/Assets/Scripts/Biome/Generator/AssortedWorldGenerator.cs
@@ -1,8 +1,8 @@
-﻿using UnityEngine;
+﻿using Assets.Scripts.World.Chunk;
+using UnityEngine;
 
-namespace Assets.Scripts.Biomes
+namespace Assets.Scripts.Biome.Generator
 {
-
     [CreateAssetMenu(fileName = "Generator", menuName = "Blockycraft/Generators/Assorted")]
     public sealed class AssortedWorldGenerator : WorldGenerator
     {
