+++ blockycraft/Assets/Scripts/Geometry/BlockFace.cs
@@ -1,9 +1,12 @@
-public enum BlockFace
+namespace Assets.Scripts.Geometry
 {
-    Left = 0,
-    Right = 1,
-    Top = 2,
-    Bottom = 3,
-    Front = 4,
-    Back = 5,
+    public enum BlockFace
+    {
+        Left = 0,
+        Right = 1,
+        Top = 2,
+        Bottom = 3,
+        Front = 4,
+        Back = 5,
+    }
 }
