+++ blockycraft/Assets/Scripts/Geometry/BlockType.cs
@@ -1,50 +1,54 @@
 ﻿using UnityEngine;
-using Assets.Scripts.Biomes;
 using System;
+using Assets.Scripts;
 
-[CreateAssetMenu(fileName = "BlockType", menuName = "Blockycraft/Block Type")]
-public sealed class BlockType : ScriptableObject
+namespace Assets.Scripts.Geometry
 {
-
-    [Header("Descriptors")]
-    public string blockName;
-
-    public Material material;
-    public TexturePack textures;
-
-    [Header("Properties")]
-    public bool isVisible;
-
-    [Header("Texture Faces")]
-    public string left;
-    public string right;
-    public string top;
-    public string bottom;
-    public string front;
-    public string back;
-
-    public BlockType() {
-        isVisible = true;
-    }
-
-    public static TexturePack.Element GetTextureID(BlockType block, int index)
+    [CreateAssetMenu(fileName = "BlockType", menuName = "Blockycraft/Block Type")]
+    public sealed class BlockType : ScriptableObject
     {
-        switch ((BlockFace)index)
+
+        [Header("Descriptors")]
+        public string blockName;
+
+        public Material material;
+        public TexturePack textures;
+
+        [Header("Properties")]
+        public bool isVisible;
+
+        [Header("Texture Faces")]
+        public string left;
+        public string right;
+        public string top;
+        public string bottom;
+        public string front;
+        public string back;
+
+        public BlockType()
         {
-            case BlockFace.Back:
-                return block.textures.Find(block.back);
-            case BlockFace.Front:
-                return block.textures.Find(block.front);
-            case BlockFace.Top:
-                return block.textures.Find(block.top);
-            case BlockFace.Bottom:
-                return block.textures.Find(block.bottom);
-            case BlockFace.Left:
-                return block.textures.Find(block.left);
-            case BlockFace.Right:
-                return block.textures.Find(block.right);
-            default:
-                throw new NotSupportedException();
+            isVisible = true;
+        }
+
+        public static TexturePack.Element GetTextureID(BlockType block, int index)
+        {
+            switch ((BlockFace)index)
+            {
+                case BlockFace.Back:
+                    return block.textures.Find(block.back);
+                case BlockFace.Front:
+                    return block.textures.Find(block.front);
+                case BlockFace.Top:
+                    return block.textures.Find(block.top);
+                case BlockFace.Bottom:
+                    return block.textures.Find(block.bottom);
+                case BlockFace.Left:
+                    return block.textures.Find(block.left);
+                case BlockFace.Right:
+                    return block.textures.Find(block.right);
+                default:
+                    throw new NotSupportedException();
+            }
         }
     }
 }
