+++ blockycraft/Assets/Scripts/Geometry/Voxel.cs
@@ -1,13 +1,15 @@
 ﻿using UnityEngine;
 
-public static class Voxel
+namespace Assets.Scripts.Geometry
 {
-    public const int VerticesInFace = 4;
-    public const int NumberOfFaces = 6;
-
-    public static readonly Vector3 Center = new Vector3(0.5f, 0.5f, 0.5f);
-    public static readonly Vector3[] Vertices = new Vector3[8]
+    public static class Voxel
     {
+        public const int VerticesInFace = 4;
+        public const int NumberOfFaces = 6;
+
+        public static readonly Vector3 Center = new Vector3(0.5f, 0.5f, 0.5f);
+        public static readonly Vector3[] Vertices = new Vector3[8]
+        {
         new Vector3(0.0f, 0.0f, 0.0f),
         new Vector3(1.0f, 0.0f, 0.0f),
         new Vector3(1.0f, 1.0f, 0.0f),
@@ -16,13 +18,13 @@ public static class Voxel
         new Vector3(1.0f, 0.0f, 1.0f),
         new Vector3(1.0f, 1.0f, 1.0f),
         new Vector3(0.0f, 1.0f, 1.0f)
-    };
+        };
 
-    public static readonly int[] Triangles = new int[6]{
+        public static readonly int[] Triangles = new int[6]{
         0,1,2,2,1,3
     };
-    
-    public static readonly int[,] Tris = new int[NumberOfFaces, VerticesInFace] {
+
+        public static readonly int[,] Tris = new int[NumberOfFaces, VerticesInFace] {
         {0, 3, 1, 2},
         {5, 6, 4, 7},
         {3, 7, 2, 6},
@@ -31,11 +33,11 @@ public static class Voxel
         {1, 2, 5, 6}
     };
 
-    public static float Scale = 1.0f;
+        public static readonly float Scale = 1.0f;
 
-    public static Vector3 Position(Vector3Int coord)
-    {
-        //(coord.x * Vector3.left * Scale) + (coord.y * Vector3.up * Scale) + (coord.z * Vector3.forward * Scale);
-        return new Vector3(-1 * coord.x * 1.0f  * Scale, coord.y * Scale, coord.z * Scale); 
+        public static Vector3 Position(Vector3Int coord)
+        {
+            return new Vector3(-1 * coord.x * 1.0f * Scale, coord.y * Scale, coord.z * Scale);
+        }
     }
 }
