+++ blockycraft/Assets/Scripts/TexturePack.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Assets.Scripts.Biomes
+namespace Assets.Scripts
 {
     [CreateAssetMenu(fileName = "Pack", menuName = "Blockycraft/Texture Pack")]
     public sealed class TexturePack : ScriptableObject
@@ -14,7 +14,8 @@ namespace Assets.Scripts.Biomes
         private System.Collections.Generic.Dictionary<string, Element> lookup;
         public string defaultKey;
 
-        public Element Find(string key) {
+        public Element Find(string key)
+        {
             if (lookup == null)
             {
                 lookup = new System.Collections.Generic.Dictionary<string, Element>();
@@ -35,7 +36,8 @@ namespace Assets.Scripts.Biomes
             );
         }
 
-        public Vector2 Dimensions(Element element) {
+        public Vector2 Dimensions(Element element)
+        {
             return new Vector2(
                 (float)element.Width / Width,
                 -(float)element.Height / Height
